+++ src/algorithm/dp/rubberHouse/rubberHouseEasy.ts
@@ -1,9 +1,9 @@
-export function rubberHouseEasy(numbers: number[], start = 0, end = numbers.length - 1): number {
+export function rubberHouseEasy(numbers: number[], startIndex = 0, endIndex = numbers.length - 1): number {
   let first = 0;
   let second = 0;
   let current = 0;
 
-  for (let i = start; i <= end; i++) {
+  for (let i = startIndex; i <= endIndex; i++) {
     current = Math.max(first + numbers[i], second);
     first = second;
     second = current;